same mock interface as web interface
